returns the number of bits you would need to flip to get from one number to another.

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned long int or;
-	unsigned long int retvalue;
+	unsigned int retvalue;
 
 	or = n ^ m;
